feat: refactor code, add description in readme
