+++ src/files/moveFile.js
@@ -1,4 +1,4 @@
-import { operationFailedMassage, successMassage } from '../massages/messages.js';
+import { operationFailedMassage } from '../massages/messages.js';
 import { copyFile } from './copyFile.js';
 import { unlink } from 'fs/promises';
 import { getFilePath } from '../pathManager/getFilePath.js';
@@ -8,7 +8,6 @@ export const moveFile = async (prevPath, pathToF, pathToNewDirectory) => {
     await copyFile(prevPath, pathToF, pathToNewDirectory);
     const pathToFile = getFilePath(prevPath, pathToF);
     await unlink(pathToFile);
-    console.log(successMassage());
   } catch(error) {
     console.error(operationFailedMassage(), error);
   }
