+++ src/os/osFunctions.js
@@ -1,20 +1,26 @@
 import * as os from 'node:os';
+import { operationFailedMassage } from '../massages/messages.js';
 
 export const getOsEOL = () => {
   console.log(JSON.stringify(os.EOL));
 };
 
 export const getOsCpus = () => {
-  const cpusArr = os.cpus().map((elem) => {
-    return {
-      model: elem.model,
-      rate: elem.speed
-  }});
-  const finalObj = {
-    amount: os.cpus().length,
-    cpus: cpusArr 
-    }
-  console.log(finalObj);
+  try {
+    const cpusArr = os.cpus().map((elem) => {
+      const speed = elem.speed / 1000 < 0.1 ? elem.speed * 100 : elem.speed;
+      return {
+        model: elem.model,
+        rate:  speed
+    }});
+    const finalObj = {
+      amount: os.cpus().length,
+      cpus: cpusArr 
+      }
+    console.log(finalObj);
+  } catch(error) {
+    console.error(operationFailedMassage(), error)
+ }
 }
 
 export const getOsArch = () => {
